requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
pdf2docx>=0.5.8
Pillow>=10.0.0
img2pdf>=0.5.1
python-magic-bin>=0.4.14
